LATIR-14: Rebasing and minor changes
